requirements.txt
argcomplete==3.0.8
asgiref==3.6.0
attrs==23.1.0
black==23.3.0
boto3==1.26.132
botocore==1.29.132
certifi==2023.5.7
cffi==1.15.1
cfn-flip==1.3.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==3.4.8
Deprecated==1.2.13
Django==4.1.7
django-cors-headers==3.14.0
django-oauth-toolkit==2.2.0
django-s3-storage==0.14.0
djangorestframework==3.14.0
drf-spectacular==0.26.2
durationpy==0.5
exceptiongroup==1.1.1
flake8==6.0.0
hjson==3.1.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
isort==5.12.0
jmespath==1.0.1
jsonschema==4.17.3
jwcrypto==1.4.2
kappa==0.6.0
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.1
placebo==0.9.0
platformdirs==3.5.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
pyrsistent==0.19.3
pytest==7.3.1
pytest-django==4.5.2
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0
requests==2.30.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tqdm==4.65.0
troposphere==4.3.2
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
Werkzeug==2.3.4
wrapt==1.15.0
zappa==0.56.1
zappa-django-utils==0.4.1
